backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
requests>=2.31.0
pytz>=2023.3
cryptography>=41.0.0
python-dotenv>=1.0.0
websockets>=12.0
pytest>=7.4.0
pytest-mock>=3.11.1
